Add PAGES_URL environment variable for deployment
This commit adds the `PAGES_URL` environment variable to the GitHub Actions workflow. The value of this variable is set to "http://${{ github.repository_owner }}.github.io/${{ github.repository }}". Additionally, the `yarn build` command in the workflow now includes the `--base "${{ env.PAGES_URL }}"` flag for building the project with the correct base URL.