Add broadcast functionality to service worker
This commit adds the necessary code to enable broadcasting messages from the service worker to clients. A new BroadcastMessage interface is introduced, and a BroadcastChannel is created with the channel name "service_worker_channel". When an output event occurs, a message containing the output data is sent through the channel.