Add support for controlling the service worker
This commit adds support for controlling the service worker by handling commands received through a broadcast channel. The `handleCommand` function is responsible for processing the commands and taking appropriate actions based on the command content. The supported commands include "start", "pause", and "stop". When a "start" command is received, the `boardRunner` instance is created if it doesn't exist and then started. Similarly, when a "pause" or "stop" command is received, the corresponding methods of `boardRunner` are called to pause or stop the execution of the board.