Update FormComponent to include placeholder for input fields
The FormComponent in src/components/FormComponent.tsx has been updated to include a placeholder attribute for the input fields. This will provide a visual hint or example value for the user when filling out the form.

refactor: Modify service worker code and add new schema

The service worker code in src/service-worker/sw.ts has been modified. The commented out section that creates inputs and wires them to outputs has been removed. Instead, new inputs and wires have been added using a different approach.

Additionally, a new schema called "schema2" has been defined with a single property "message_2". This schema is used to create an input and wire it to an output in the board object.