Add dismissInputOnSend function
This commit adds a new function called `dismissInputOnSend` to the `InputRequestsRenderer.tsx` file. This function is used to handle the dismissal of input requests when a response is sent. It takes in two parameters, `setRequests` and `request`, and returns a callback function that removes the request from the list of messages.