Add headless.ts file for running board in a headless environment
This commit adds a new file, `headless.ts`, which allows the board to be run in a headless environment. The script uses the `LogProbe` from the `@google-labs/breadboard` package and sets up the necessary inputs for running the board. If an input of type "output" is received, it logs the node ID and outputs to the console.